ais/requirements/requirements.txt
affine>=2.4,<3
aiohttp>=3.9,<4
aiosignal>=1.3,<2
appdirs>=1.4,<2
attrs>=23.2,<24
bleach>=6.1,<7
bokeh>=3.3,<4
cachetools>=5.3.3,<6
cartopy>=0.23,<1
certifi>=2023.11,<2024.7
cftime>=1.6,<2
charset-normalizer>=3.3,<4
click>=8.1,<9
click-plugins>=1.1,<2
cligj>=0.7,<1
cloudpickle>=3.0,<4
colorcet>=3.0,<4
contextily>=1.5,<2
contourpy>=1.2,<2
cycler>=0.12,<1
dask>=2023.4,<2025
dask[distributed]>=2023.4,<2025
datashader>=0.16,<1
decorator>=5.1,<6
docker-pycreds>=0.4,<1
fastavro>=1.9,<2
filelock>=3.13,<4
fiona>=1.9,<2
fonttools>=4.47,<5
frozenlist>=1.4,<2
fsspec>=2023.6,<2024
fvcore>=0.1,<1
gcsfs>=2023.6,<2024
geographiclib>=2.0,<3
geopandas>=0.14,<1
geopy>=2.4,<3
gitdb>=4.0,<5
gitpython>=3.1,<4
google-auth==2.*,<3
google-cloud-storage==2.*,<3
grpcio>1.65.1,<2
grpcio-health-checking>=1,<2
holoviews>=1.18,<2
huggingface-hub>=0.20,<1
idna>=3.6,<4
importlib-metadata>=7.0,<8
interrogate>=1.7,<2
iopath>=0.1,<1
jinja2>=3.1,<4
joblib>=1.3,<2
kiwisolver>=1.4,<2
linkify-it-py>=2.0,<3
llvmlite>=0.41,<1
locket>=1.0,<2
markdown>=3.5,<4
markdown-it-py>=3.0,<4
markupsafe>=2.1,<3
matplotlib>=3.8,<4
mdit-py-plugins>=0.4,<1
mdurl>=0.1,<1
mercantile>=1.2,<2
mpmath>=1.3,<2
multidict>=6.0,<7
multimethod>=1.10,<2
multipledispatch>=1.0,<2
mypy-extensions>=1.0,<2
netcdf4>=1.6,<2
networkx>=3.2,<4
numba>=0.58,<1
numpy>=1.26,<2
numpy-indexed>=0.3,<1
oauthlib>=3.2,<4
packaging>=23.2,<24
pandas>=2.2,<3
pandera>=0.19,<1
panel>=1.3,<2
param>=2.0,<3
partd>=1.4,<2
peft==0.6.0
pillow>=9.5,<10
portalocker>=2.8,<3
protobuf>=4.25,<5
psutil>=5.9,<6
pyarrow>=15.0,<16
pyasn1>=0.5,<1
pyasn1-modules>=0.3,<1
pycocotools>=2.0,<3
pyct>=0.5,<1
pydantic>=2.7,<3
pyparsing>=3.1,<4
pyproj>=3.6,<4
pytest>7,<8
pytest-xdist
python-dateutil>=2.8,<3
pytz
pyviz-comms>=3.0,<4
pyyaml>=6.0,<7
ray[serve]==2.35.0  # This is pinned hard since we need it to match what is in atlas-ray.tf. If you change this, you must also change that.
rdp>=0.8,<1
requests>=2.31,<3
requests-oauthlib>=1.3,<2
retrying>=1.3,<2
rsa>=4.9,<5
scikit-learn>=1.3,<2
scipy>=1.12,<2
seaborn>=0.12,<1
sentry-sdk>=1.39,<2
setproctitle>=1.3,<2
shapely>=2.0,<3
six>=1.16,<2
smmap>=5.0,<6
snuggs>=1.4,<2
spatialpandas>=0.4,<1
suncalc>=0.1.2,<1.0
sympy>=1.12,<2
tabulate>=0.9,<1
termcolor>=2.4,<3
threadpoolctl>=3.2,<4
toolz>=0.12,<1
torch==2.1.2
tornado>=6.4,<7
tqdm>=4.66,<5
typeguard>=4.1,<5
typing-extensions>=4.9,<5
typing-inspect>=0.9,<1
tzdata>=2023.4
uc-micro-py>=1.0,<2
urllib3>=1.26,<2
wandb>=0.16,<1
webencodings>=0.5,<1
wrapt>=1.16,<2
xarray>=2024.1,<2025
xyzservices>=2023.10,<2024
yacs>=0.1,<1
yarl>=1.9,<2
zipp>=3.17,<4
